Added file bind extension methods
