+++ src/Extensions/RouteDataExtensions.cs
@@ -1,4 +1,4 @@
-namespace Botwin
+namespace Botwin.Extensions
 {
     using System;
     using Microsoft.AspNetCore.Routing;
