+++ test/BindTests.cs
@@ -8,6 +8,7 @@ namespace Botwin.Tests
     using System.Reflection;
     using System.Text;
     using System.Threading.Tasks;
+    using Botwin.Extensions;
     using FluentValidation;
     using FluentValidation.Results;
     using Microsoft.AspNetCore.Hosting;
